sllm_qa_chatbot_prototype/requirements.txt
langchain
langchain_community
python-dotenv
pandas
pytest
